requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
croniter==2.0.5
cryptography==39.0.0
dj-database-url==0.5.0
Django==3.2.7
django-auditlog==2.1.1
django-autocomplete-light==3.8.2
django-axes==5.25.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-formtools==2.3
django-geojson==3.2.0
django-hijack==3.0.2
django-ipware==4.0.0
django-js-asset==2.2.0
django-leaflet==0.28.2
django-log-viewer==1.1.4
django-mathfilters==1.0.0
django-menu-generator==1.1.0
django-multi-form-view==2.0.1
django-multiselectfield==0.1.12
django-qsessions==1.1.2
django-ranged-response==0.2.0
django-redis==5.0.0
django-rq==2.4.1
django-session-security==2.6.6
django-simple-captcha==0.5.14
django-waffle==2.2.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.15.0
djangorestframework-datatables==0.7.1
djongo==1.3.6
dnspython==2.1.0
emoji==1.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
mysqlclient==2.2.1
ncclient==0.6.13
netaddr==0.8.0
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.1
paramiko==3.4.0
passlib==1.7.4
Pillow==8.3.2
platformdirs==4.2.1
ply==3.11
pyasn1==0.4.8
pyclean==3.0.0
pycparser==2.22
pycryptodomex==3.20.0
pyminizip==0.2.6
pymongo==3.12.3
PyNaCl==1.5.0
pyrad==2.4
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-redis-lock==3.7.0
pytz==2024.1
redis==5.0.4
requests==2.26.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rq==1.10.0
rq-scheduler==0.11.0
six==1.16.0
soupsieve==2.5
spyne==2.14.0
sqlparse==0.2.4
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==1.26.18
user-agents==2.2.0
whitenoise==6.6.0
wkhtmltopdf==0.2
xlrd==2.0.1
XlsxWriter==3.1.6
xmltodict==0.13.0
zeep==4.1.0
